pyproject.toml
[project]
name = "mcp-server-financial-analyzer"
version = "0.1.0"
description = "MCP server for financial analysis - stock prices, fundamentals, and SEC filings"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edgartools>=5.28.4",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "yfinance>=1.2.0",
]

[project.scripts]
financial-analyzer = "financial_analyzer.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/financial_analyzer"]
